Listing the gists.
Now the script has the functionallity to list yours (or others) gists.

It reads from '~/.gists.rc' configuration file.

It has the --private (-p) argument to retrieve the private gists as
well.

It has the --user (-u) argument to override the 'user' property in
'gists.rc'

It has the --secret (-s) argument to override the 'secret' property in
'gists.rc'